Fix database and fix sky levels
git-svn-id: svn+ssh://astrometry.net/svn/trunk/projects/tractor@21578 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ allNSAtlas.py
@@ -7,20 +7,26 @@ import numpy as np
 import pylab as plt
 import pyfits
 import sys
+import traceback
 
 from general import generalNSAtlas
 from halflight import halflight
 #from addtodb import add_to_table_nsatlas
 
 def main():
-    #nsgals = [167,20315,24188,39462,44000,131926,133189,133385,133469,133678,140045,149936,151192,151193,152228,158585]
-    nsgals2 = [39462,131926,140045,149936,151192,151193,152228,158585]
+    nsgals2=[]
+    f = open("targets.txt","r")
+    for line in f:
+        nsgals2.append(int(line))
 
-    #133678 seems not to be in the table???
+    print nsgals2
     for entry in nsgals2:
         try:
             print entry
             newentry = 'NSA_ID_%s' % entry
+            if os.path.exists("NSAtlas_Output/%s.pickle" % newentry):
+                print "%s has run already!" % newentry
+                continue
             print 'running tractor for %s' %entry
             generalNSAtlas(entry,itune1=6,itune2=6,nocache=True)
             os.system('cp flip-%s.pdf NSAtlas_Output' % newentry)
@@ -28,7 +34,13 @@ def main():
             os.system('cp %s.pickle NSAtlas_Output' % newentry)
             halflight(newentry,direc='NSAtlas_Output')
         except AssertionError:
-            print 'Tractor failed on %s' %entry
+            print traceback.print_exc()
+            print "Tractor failed on entry: %s" % newentry
+            continue
+        except IndexError:
+            print traceback.print_exc()
+            print "Tractor failed on entry: %s" % newentry
+            continue
         #add_to_table_nsatlas(newentry)
 
 if __name__ == '__main__':
